proposed quick fix to checker.yaml
This is a crappy way to do this, but I don't know how to fix the issue that the layout isn't displaying the field labelled "title" which explains you should enter a zone name, so I just prefixed the prompt for the zone name with the text in the 'title' field. I won't be offended if you ignore this, because I am not sure it will help users who are confused anyway, but at least I took a stab at it. I did change the English version slightly, but did not feel qualified to muck with the other languages. Maybe it would be better to give some example zone names (that pass the test)??