added text to clarify goals; revised other text for clarity and grammar
this started out as an attempt to add text clarifing that the proposed
EDNS buffer sizes are not meant as a permanent change to the EDNS
specification, but an agreed-upon default configuration, intended for
use in the absence of better information about MTU, which will prevent
fragmentation on the majority of present-day networks.

in the process, I ended up making a lot of other revisions for clarity
and grammar, and changed the references to 1220 bytes to 1232, since
that seems to be the growing consensus in the slack group.

(I should probably have broken this into multiple commits but didn't
think of it in time, sorry.)